06_flask_fastapi/queimadas_fastapi/pyproject.toml
[project]
name = "queimadas-fastapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.2.2",
    "databricks-sql-connector[pyarrow]>=4.2.5",
]
